[218] libdh/math: some retypes of finctions

+++ include/dhmath.h
@@ -16,7 +16,7 @@ extern long double sqrt_fld_dh(long double a);
 
 extern double mad_d_dh(double a);
 
-extern long double quadequi_dh(long double a, long double b, long double c, int x);
+extern double quadequi_dh(double a, double b, double c, int x);
 
 extern long double sin_deg_dh(long double a);
